[sway] make move-output-to-ws more resilient
it started crashing when an output was disabled, because sway goesn't
give us an `id` field for those displays.  we weren't use id anyway, but
we probably don't wanna move stuff to disabled displays, so we filter
those out.

also, if you escaped out of the tofi prompt, the script would error.
now it detects a missing choice and just silently does nothing.